example: a conforming container class and library
The triplet.js class implmentes a subset of the Array
interface and uses the isArray spec to signal that is
should be treated an array.

The examplelib in examplelib-conforms has two methods
which use two different array checking techniques on their
arguement. Both also conform to the isArray spec.

The conforms.html has a small snippet to test this funcitonality.
First calling the examplelib on native JavaScript arrays, and
then calling them on triplet objects.