example: a non-conforming library working with a polyfill
The examplelib in examplelib-compatible has two methods
which use two different array checking techniques on their
arguement. Both use JavaScript best practices, but neither
conforms to the isArray spec, and so as written they will
only work on native JavaScript array objects.

The isArray-polyfill patches the Object.prototype.toString
and Array.isArray functions so that they themselves return
results that conform to the isArray spec.

The polyfill.html has a small snippet to test this funcitonality.
First calling the examplelib on native JavaScript arrays, and
then calling them on triplet objects, which conform to the
isArray spec and advertise their ability to implmenet
JavaScript array functions.